+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and 
 
 ## [Unreleased]
 
+## [4.0.0] — 2024–08–14
+
 ### Changed
 
 - The minimum required version of `stylelint` has been increased to `16.8.1`.
@@ -44,7 +46,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and 
 
 - Base config for `stylelint`.
 
-[Unreleased]: https://github.com/firefoxic/stylelint-config/compare/v3.0.0...HEAD
+[Unreleased]: https://github.com/firefoxic/stylelint-config/compare/v4.0.0...HEAD
+[4.0.0]: https://github.com/firefoxic/stylelint-config/compare/v3.0.0...v4.0.0
 [3.0.0]: https://github.com/firefoxic/stylelint-config/compare/v2.0.0...v3.0.0
 [2.0.0]: https://github.com/firefoxic/stylelint-config/compare/v1.0.0...v2.0.0
 [1.0.0]: https://github.com/firefoxic/stylelint-config/releases/tag/v1.0.0